+++ TimeSeriesDL/model/__init__.py
@@ -2,5 +2,6 @@
 from .conv_lstm_model import ConvLSTM
 from .auto_encoder.conv_ae_model import ConvAE
 from .auto_encoder.conv_vae import ConvVAE
+from .auto_encoder.dummy_ae import DummyAutoEncoder
 from .base_model import BaseModel
 from .lstm_model import LSTM
